whisper_server/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
openai-whisper>=20231117  # OpenAI's Whisper package
numpy>=1.21.0
python-multipart  # 用于处理文件上传
python-dotenv>=0.19.0  # 用于加载.env配置文件
